Update blog page styles
