+++ package.json
@@ -81,7 +81,7 @@
     "sonner": "^1.3.1",
     "tailwind-merge": "^1.14.0",
     "tailwindcss-animate": "^1.0.7",
-    "three": "^0.158.0",
+    "three": "^0.159.0",
     "vaul": "^0.9.1",
     "zod": "^3.22.4"
   },
